Create functional test group and fixup some code review feedback
