+++ .travis.yml
@@ -23,6 +23,9 @@ matrix:
           env: DEBUGGER_TEST_RELEASE=true
         - os: linux
           python: "3.6-dev"
+          env: FUNCTIONAL_TEST=true
+        - os: linux
+          python: "3.6-dev"
           env: SINGLE_WORKSPACE_TEST=true
         - os: linux
           python: "3.6-dev"
@@ -36,6 +39,9 @@ matrix:
         - os: linux
           python: "3.7-dev"
           env: DEBUGGER_TEST_RELEASE=true
+        - os: linux
+          python: "3.7-dev"
+          env: FUNCTIONAL_TEST=true
         - os: linux
           python: "3.7-dev"
           env: SINGLE_WORKSPACE_TEST=true
@@ -82,6 +88,10 @@ script:
   - npm run debugger-coverage
   - if [ $TRAVIS_UPLOAD_COVERAGE == "true" ]; then
         bash <(curl -s https://codecov.io/bash);
+    fi
+  - if [ $FUNCTIONAL_TEST == "true" ]; then
+        npm run cover:enable;
+        npm run test:functional;
     fi
   - if [ $SINGLE_WORKSPACE_TEST == "true" ]; then
         npm run cover:enable;
