+++ src/client/datascience/jupyterAvailability.ts
@@ -4,12 +4,15 @@
 import { inject, injectable } from 'inversify';
 
 import { IPythonExecutionFactory } from '../common/process/types';
+import { ILogger } from '../common/types';
 import { IJupyterAvailability } from './types';
 
 @injectable()
 export class JupyterAvailability implements IJupyterAvailability {
 
-    constructor(@inject(IPythonExecutionFactory) private executionFactory: IPythonExecutionFactory) {
+    constructor(
+        @inject(IPythonExecutionFactory) private executionFactory: IPythonExecutionFactory,
+        @inject(ILogger) private logger : ILogger) {
 
     }
 
@@ -19,7 +22,8 @@ export class JupyterAvailability implements IJupyterAvailability {
             const pythonService = await this.executionFactory.create({});
             const result = await pythonService.execModule('jupyter', ['notebook', '--version'], { throwOnStdErr: true, encoding: 'utf8' });
             return (!result.stderr);
-        } catch {
+        } catch (err) {
+            this.logger.logWarning(err);
             return false;
         }
     }
@@ -30,7 +34,8 @@ export class JupyterAvailability implements IJupyterAvailability {
             const pythonService = await this.executionFactory.create({});
             const result = await pythonService.execModule('jupyter', ['nbconvert', '--version'], { throwOnStdErr: true, encoding: 'utf8' });
             return (!result.stderr);
-        } catch {
+        } catch (err) {
+            this.logger.logWarning(err);
             return false;
         }
     }
