+++ src/test/common.ts
@@ -305,22 +305,3 @@ export async function isPythonVersion(...versions: string[]): Promise<boolean> {
 export interface IExtensionTestApi extends IExtensionApi {
     serviceContainer: IServiceContainer;
 }
-// Custom module loader so we skip .css files that break non webpack wrapped compiles
-// tslint:disable-next-line:no-var-requires no-require-imports
-const Module = require('module');
-
-// tslint:disable-next-line:no-function-expression
-(function () {
-    const origRequire = Module.prototype.require;
-    const _require = (context, filepath) => {
-        return origRequire.call(context, filepath);
-    };
-
-    Module.prototype.require = function (filepath) {
-        if (filepath.endsWith('.css')) {
-            return '';
-        }
-        // tslint:disable-next-line:no-invalid-this
-        return _require(this, filepath);
-    };
-})();
